Made "Buy Akita Inu Now" buttons on home page point to /how-to-buy
These previously pointed to Tinyman.  Problem with this is that
people who are unfamiliar with the Algorand ecosystem, but who want
to buy Akita Inu, will be completely lost if they are dropped on
Tinyman.  Many will just give up.  It makes much more sense to take
them to the /how-to-buy page that will walk them through the whole
buying process (of which swapping on Tinyman is just the last step).

People who already hold Algo and know how to use Tinyman will still
be able to find Tinyman from the /how-to-buy page, and I have left
the "Swap on Tinyman" header/footer links there too for the moment.
People who are new to the Algorand ecosystem will probably just
ignore the "Swap on Tinyman" buttons, cos they won't know what that
means.

So basically my thinking is that "Buy Now" links should lead to
/how-to-buy.  Links to Tinyman should say "Swap on Tinyman" or
similar.