Conversion Improvements to Master (#18)
* Create README.md

created readme, but there's not much to read.

* How to buy page

How to buy page with content and icons from google drive documents.

Version 1.

* Styling changes, navbar size bug fix, and how to buy page

* AKITA INU -> AKITA INU TOKEN

* Made "Buy Akita Inu Now" buttons on home page point to /how-to-buy

These previously pointed to Tinyman.  Problem with this is that
people who are unfamiliar with the Algorand ecosystem, but who want
to buy Akita Inu, will be completely lost if they are dropped on
Tinyman.  Many will just give up.  It makes much more sense to take
them to the /how-to-buy page that will walk them through the whole
buying process (of which swapping on Tinyman is just the last step).

People who already hold Algo and know how to use Tinyman will still
be able to find Tinyman from the /how-to-buy page, and I have left
the "Swap on Tinyman" header/footer links there too for the moment.
People who are new to the Algorand ecosystem will probably just
ignore the "Swap on Tinyman" buttons, cos they won't know what that
means.

So basically my thinking is that "Buy Now" links should lead to
/how-to-buy.  Links to Tinyman should say "Swap on Tinyman" or
similar.

* A number of small edits to /how-to-buy

- Paragraph at top to hopefully avoid confusion with the ERC-20
  token.  Otherwise there's a risk people ignore our 6 steps and buy
  the wrong token on their exchange.
- Terminology: Algorand the blockchain, ALGO the currency; Akita Inu
  ASA is a token not a coin.
- Removed * that didn't belong
- Added extra links to Tinyman to help people get to the right place
- Changed title of step 6 to make it clearer that it's a swap.
- Made it clear in step 6 that the Akita will arrive in the wallet
  quickly and automatically.  I've seen people ask about that.
- Removed bit saying you can swap other ASAs too, cos these
  instructions are not really for people who already hold ASAs.
- Other minor edits to try to make the process a little clearer.

Sure it could be improved further, if anyone wants to change more.

* Added a footer link to /how-to-buy.

It's a really important page, we should have links to it all over
the place.

* Changed title of /how-to-buy page.

Co-authored-by: maquino1985 <maquino1985@gmail.com>
Co-authored-by: AlgoSlurp <trippymonk55@protonmail.com>
Co-authored-by: SomeEnormousBear <bear@speedymail.org>