Fix for Issue 20 - internal linking.
Now we have extensionless URLs in the HTML, and links to "/"
instead of "/index.html", which is the way the links should be for
the HTML on the live website.  No web server is required for local
testing (you can still just open the HTML in a browser and click
between pages) because I added some JavaScript that will only run
when the website is being tested locally.  Sort of like a basic
web server, but nobody needs to install anything to test and
modify the site, keeping it really simple for new people starting
work on the website.